MO-1980 POM offboarding bulk reallocation
Ticket: https://dsdmoj.atlassian.net/browse/MO-1980

New entry-point to the bulk reallocation. Upon marking a POM as "No longer at $prison" (internal status `deleted`), we present confirmation and, then enter the bulk reallocation journey if they had any primary cases still allocated.

If SPO drops from the journey and POM still has primary cases allocated, they can re-enter the bulk reallocation fro the "Attention needed" staff tab.

Lots of specs refactor/cleanup now that we have an almost finalised feature and we can link together some previously isolated journeys.

+++ app/controllers/poms_controller.rb
@@ -3,9 +3,8 @@
 class PomsController < PrisonStaffApplicationController
   before_action :ensure_spo_user
 
-  before_action :load_pom_staff_member, only: [:show, :edit, :update, :reallocate, :confirm_removal, :destroy]
-  before_action :store_referrer_in_session, only: [:edit]
-  before_action :set_referrer
+  before_action :load_pom_staff_member, except: [:index, :show_non_pom]
+  before_action :ensure_pom_is_editable, only: [:edit, :update]
 
   def index
     @poms = @prison.get_list_of_poms.sort_by(&:full_name_ordered)
@@ -50,6 +49,10 @@ class PomsController < PrisonStaffApplicationController
     @edit_pom_form = PomProfileForm.new(edit_pom_params.to_h)
 
     if @edit_pom_form.valid?
+      if @edit_pom_form.deleting?
+        redirect_to confirm_delete_prison_pom_path(active_prison_id, nomis_staff_id) and return
+      end
+
       pom_detail = @prison.pom_details.find_by!(nomis_staff_id:)
       pom_detail.working_pattern = @edit_pom_form.working_pattern_ratio
       pom_detail.status = @edit_pom_form.status
@@ -83,13 +86,51 @@ class PomsController < PrisonStaffApplicationController
 
   def confirm_removal; end
 
+  def confirm_delete
+    @confirm_delete_form = ConfirmDeletePomForm.new
+  end
+
   def destroy
-    NomisUserRolesService.remove_pom(@prison, nomis_staff_id)
-    redirect_to prison_poms_path(anchor: "#{params[:from]}!top"),
-                notice: "#{@pom.full_name_or_staff_id} removed. If necessary, their cases have been moved to @unallocated_link@."
+    # Legacy flow (limbo poms removal without confirmation form)
+    # TODO: to be removed once we release the new bulk-reallocation feature
+    unless params.key?(:confirm_delete_pom)
+      NomisUserRolesService.remove_pom(@prison, nomis_staff_id)
+      redirect_to prison_poms_path(anchor: 'attention_needed!top'),
+                  notice: "#{@pom.full_name_or_staff_id} removed. If necessary, their cases have been moved to @unallocated_link@."
+      return
+    end
+
+    # New flow (confirm_delete form submission)
+    @confirm_delete_form = ConfirmDeletePomForm.new(confirm_delete_params)
+
+    if @confirm_delete_form.valid?
+      if @confirm_delete_form.confirmed?
+        pom_detail = @prison.pom_details.find_by!(nomis_staff_id:)
+
+        if pom_detail.has_primary_allocations?
+          pom_detail.deleted!
+          redirect_to reallocate_prison_pom_path(active_prison_id, nomis_staff_id:)
+        else
+          NomisUserRolesService.remove_pom(@prison, nomis_staff_id)
+          redirect_to prison_poms_path(active_prison_id),
+                      notice: "#{@pom.full_name_or_staff_id} successfully removed from this service"
+        end
+      else
+        redirect_to edit_prison_pom_path(active_prison_id, nomis_staff_id)
+      end
+    else
+      render :confirm_delete
+    end
   end
 
 private
+
+  def ensure_pom_is_editable
+    unless @pom.editable?
+      redirect_to prison_pom_path(active_prison_id, id: nomis_staff_id),
+                  notice: 'This POM has been removed and their profile cannot be edited.'
+    end
+  end
 
   def load_pom_staff_member
     @pom = StaffMember.new @prison, nomis_staff_id
@@ -97,6 +138,10 @@ private
 
   def edit_pom_params
     params.require(:edit_pom).permit(:working_pattern, :status, :description)
+  end
+
+  def confirm_delete_params
+    params.fetch(:confirm_delete_pom, {}).permit(:confirmation)
   end
 
   def nomis_staff_id
